[APIS-904] Add connection error handling (#56)
http://jira.cubrid.org/browse/APIS-904

You can connect to the server using SQLDriverConnect().
The connection method can be used using DSN and connection url. At this time, if autocommit is not set, it may be null. The cause of the connection error is that autocommit, which is null, was compared with a string.
Exception handling is required when autocommit is null.